bootcamp ownership conditions

+++ controllers/bootcamps.js
@@ -111,7 +111,7 @@ exports.createBootcamp = asyncHandler(async (req,res,next) =>{
     if(publishedBootcamp && req.user.role !== 'admin'){
         return next(new ErrorResponse(`The user with ID ${req.user.id} has already published a bootcamp`, 400))
     }
-    
+
         const bootcamp = await Bootcamp.create(req.body)
         res.status(200).json({success:true,data:bootcamp})
 
@@ -122,14 +122,23 @@ exports.createBootcamp = asyncHandler(async (req,res,next) =>{
 // @route PUT /api/v1/bootcamps/:id
 // @access Private
 exports.updateBootcamp = asyncHandler(async (req,res,next) =>{
-        const bootcamp = await Bootcamp.findOneAndUpdate(req.params.id, req.body,{
-            new:true,
-            runValidator:true
-        })
+        let bootcamp = await Bootcamp.findById(req.params.id)
+
         if(!bootcamp){
             //for correctly formatted id not in db
             return res.status(400).json({success:false})
         }
+
+        // Make sure the user is bootcamp owner
+        if(bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+            return next(new ErrorResponse(`User ${req.params.id} is not authorized to update this bootcamp`,404))
+        }
+
+        bootcamp = await Bootcamp.findByIdAndUpdate(req.params.id, req.body,{
+            new:true, // return the updated bootcamp
+            runValidator:true
+        })
+
         res.status(200).json({success:true,data:bootcamp})
 
 })
@@ -143,6 +152,12 @@ exports.deleteBootcamp = asyncHandler(async (req,res,next) =>{
             //for correctly formatted id not in db
             return next(new ErrorResponse(`Bootcamp Not Found with id of ${req.params.id}`,404 ))
         }
+
+        // Make sure the user is bootcamp owner
+        if(bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+            return next(new ErrorResponse(`User ${req.params.id} is not authorized to update this bootcamp`,404))
+        }
+
         // Trigger the pre-hook middleware to delete associated courses
         await bootcamp.deleteOne();
         res.status(200).json({success:true,data:{}})
@@ -188,6 +203,12 @@ exports.bootcampPhotoUpload = asyncHandler(async (req,res,next) =>{
         //for correctly formatted id not in db
         return next(new ErrorResponse(`Bootcamp Not Found with id of ${req.params.id}`,404 ))
     }
+
+    // Make sure the user is bootcamp owner
+    if(bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+        return next(new ErrorResponse(`User ${req.params.id} is not authorized to update this bootcamp`,404))
+    }
+    
     // Trigger the pre-hook middleware to delete associated courses
 
     if(!req.files) {
